File tree refactoring snapshot
