+++ AngouriMath/Core/Entity/Continuous/Number/Integer.cs
