+++ AngouriMath/Core/Entity/Continuous/Number/Operators.cs
