+++ AngouriMath/Core/Entity/Continuous/Number/Rational.cs
