+++ AngouriMath/Core/Entity/Continuous/Number/Real.cs
