+++ AngouriMath/Core/Entity/Omni/Sets/PieceFunctions/Intersection.cs
