+++ AngouriMath/Core/Entity/Omni/Sets/PieceFunctions/Inversion.cs
