+++ AngouriMath/Core/Entity/Omni/Sets/PieceFunctions/Union.cs
