+++ AngouriMath/Core/Entity/Omni/Sets/SetFunctions/Intersection.cs
