+++ AngouriMath/Core/Entity/Omni/Sets/SetFunctions/Inversion.cs
