+++ AngouriMath/Core/Entity/Omni/Sets/SetFunctions/Union.cs
